+++ gbalink/main.cpp
@@ -3,7 +3,7 @@
 #include <memory>
 #include <thread>
 #include <queue>
-#include "optional.hpp"
+#include <optional>
 #include "jbus/Endpoint.hpp"
 #include "jbus/Listener.hpp"
 
@@ -72,7 +72,7 @@ CGBASupport::~CGBASupport() { SharedInstance = nullptr; }
 
 u8 CGBASupport::CalculateFusionJBusChecksum(const u8* data, size_t len) {
   u32 sum = -1;
-  for (int i = 0; i < len; ++i) {
+  for (size_t i = 0; i < len; ++i) {
     u8 ch = *data++;
     sum ^= ch;
     for (int j = 0; j < 8; ++j) {
